refactor(codegen): extract channels codegen into channels/mod.rs (Round 6)
16 fns lifted as inherent impls on Cx into channels/mod.rs:

Fallible-channel machinery (the value-error protocol):
- lower_fallible_locus_method_body
- emit_fallible_locus_method_exit_epilogue
- emit_fallible_fn_exit_epilogue
- lower_or_expr (or raise / or default / or handler / or discard
  / or fail / or substitute / or hand-off dispositions)
- lower_expr_opt (helper)
- lower_fallible_call (call-site sret plumbing)
- try_lower_fallible_stdlib_path_call (~50-arm dispatch to the
  fallible stdlib lowerings now lifted to stdlib/*.rs)
- lower_fallible_method_call
- complete_io_fallible_call, complete_parse_fallible_call
- emit_io_error_alloc, emit_parse_error_alloc
- lower_or_fail, lower_or_raise (specific dispositions)
- lower_fail (the `fail E { ... }` stmt)

Structural channel:
- resolve_failure_route (walks the parent chain to find the
  matching on_failure handler for a child-locus closure violation)

The closure-assertion lowering (lower_closure_check) shipped in
Round 4e at locus/closure.rs — it's the closure-eval fn body emit
rather than the channel dispatch, so it lived more naturally there.

Surface bumps to pub(crate):
- Cx fields: user_fns, current_user_fn_ret_alloca
- Cx methods: lower_std_os_getrandom_fallible,
  try_lower_user_locus_fallible_method,
  try_lower_form_{vec,ring_buffer,hashmap}_fallible_method,
  payload_typename_global
- FnSig struct + all fields
- Free fn: view_coerces_to

codegen.rs: 27164 -> 25333 (-1831 lines).
